Move resource intesive wfn tests
